refine decimal output format
