+++ app/src/main/java/de/moyapro/idle/domain/Biome.kt
@@ -1,7 +1,6 @@
 package de.moyapro.idle.domain
 
-import java.text.DecimalFormat
-import kotlin.math.*
+import de.moyapro.idle.util.toShortDecimalStr
 
 data class Biome(
     val name: String = "DefaultBiome",
@@ -35,19 +34,5 @@ data class Biome(
     private fun getStatusText(species: Species): String {
         return species.name + ": " + (species.getPopulationIn(this)*1E6).toShortDecimalStr() +
                 " -> " + (species.process(this.resources).getPopulation(species)*1E6).toShortDecimalStr()
-    }
-}
-
-fun Double.toShortDecimalStr() : String {
-    if (this == 0.0) return "0.0"
-    val base = floor(log10(abs(this)) / 3)
-    val outVal = round(this / 10.0.pow(base * 3) * 1E10) / 1E10
-    return when {
-        base < 1 -> outVal.toString()
-        base < 2 -> outVal.toString() + 'K'
-        base < 3 -> outVal.toString() + 'M'
-        base < 4 -> outVal.toString() + 'B'
-        base < 5 -> outVal.toString() + 'T'
-        else -> outVal.toString() + ('a' + (base.toInt()-5)/26) + ('a' + ((base.toInt()-5)%26))
     }
 }
